# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Sign in with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/accept-legal:
    post:
      operationId: acceptLegal
      tags: [auth]
      summary: Record the current user's acceptance of the legal documents
      responses:
        "200":
          description: Acceptance recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "200":
          description: Registered and signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Account already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /chat:
    post:
      operationId: sendChat
      tags: [chat]
      summary: Send a prompt and get a synchronous IQRA response
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatPromptInput"
      responses:
        "200":
          description: Chat response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatResponse"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /chat/async:
    post:
      operationId: startChatJob
      tags: [chat]
      summary: Start an asynchronous chat generation job
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatPromptInput"
      responses:
        "200":
          description: Job started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatJobStart"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /chat/async/{jobId}:
    get:
      operationId: getChatJob
      tags: [chat]
      summary: Poll an asynchronous chat generation job
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Job status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatJobStatus"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Job not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /chats:
    get:
      operationId: listChats
      tags: [chats]
      summary: List the current user's chat threads
      responses:
        "200":
          description: Chat threads
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatThreadList"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createChat
      tags: [chats]
      summary: Create a new chat thread
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatThreadInput"
      responses:
        "200":
          description: Created thread
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatThread"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /chats/{id}:
    get:
      operationId: getChat
      tags: [chats]
      summary: Get a chat thread with its messages
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Chat thread detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatThreadDetail"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Chat not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /chats/{id}/turn:
    post:
      operationId: appendChatTurn
      tags: [chats]
      summary: Append a user/assistant turn to a chat thread
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatTurnInput"
      responses:
        "200":
          description: Turn saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Chat not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    OkResponse:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean
    AuthUser:
      type: object
      required: [id, email, name, role, legalAccepted]
      properties:
        id:
          type: string
        email:
          type: string
        name:
          type: ["string", "null"]
        role:
          type: string
          enum: [user, admin]
        legalAccepted:
          type: boolean
          description: Whether the user has accepted the current version of the legal documents.
...
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    RegisterInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
        name:
          type: string
    SourceLink:
      type: object
      required: [label, href]
      properties:
        label:
          type: string
        href:
          type: string
    ChatPromptInput:
      type: object
      required: [prompt]
      properties:
        prompt:
          type: string
    ChatResponse:
      type: object
      required: [directAnswer, framework, source]
      properties:
        basmala:
          type: string
        directAnswer:
          type: string
        framework:
          type: array
          items:
            type: string
        source:
          type: string
        sourceLinks:
          type: array
          items:
            $ref: "#/components/schemas/SourceLink"
        requiresScholarReferral:
          type: boolean
        clarifyingQuestion:
          type: ["string", "null"]
        confidence:
          type: string
          enum: [high, medium, low]
        responseHeld:
          type: boolean
        structured:
          type: ["object", "null"]
          additionalProperties: true
    ChatJobStart:
      type: object
      required: [jobId, status, stage, attempt]
      properties:
        jobId:
          type: string
        status:
          type: string
          enum: [running, completed, failed]
        stage:
          type: string
        attempt:
          type: integer
    ChatJobStatus:
      type: object
      required: [status, stage, attempt]
      properties:
        jobId:
          type: string
        status:
          type: string
          enum: [running, completed, failed]
        stage:
          type: string
        attempt:
          type: integer
        lastConfidence:
          type: ["string", "null"]
          enum: [high, medium, low, null]
        error:
          type: ["string", "null"]
        basmala:
          type: string
        directAnswer:
          type: string
        framework:
          type: array
          items:
            type: string
        source:
          type: string
        sourceLinks:
          type: array
          items:
            $ref: "#/components/schemas/SourceLink"
        requiresScholarReferral:
          type: boolean
        clarifyingQuestion:
          type: ["string", "null"]
        confidence:
          type: ["string", "null"]
          enum: [high, medium, low, null]
        responseHeld:
          type: boolean
        structured:
          type: ["object", "null"]
          additionalProperties: true
    ChatThreadInput:
      type: object
      properties:
        title:
          type: string
    ChatThread:
      type: object
      required: [id, title, updatedAt]
      properties:
        id:
          type: string
        title:
          type: string
        updatedAt:
          type: string
    ChatThreadSummary:
      type: object
      required: [id, title, updatedAt, preview]
      properties:
        id:
          type: string
        title:
          type: string
        updatedAt:
          type: string
        preview:
          type: ["string", "null"]
    ChatThreadList:
      type: object
      required: [threads]
      properties:
        threads:
          type: array
          items:
            $ref: "#/components/schemas/ChatThreadSummary"
    ChatMessage:
      type: object
      required: [id, role, content, responsePayload, createdAt]
      properties:
        id:
          type: string
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
        responsePayload:
          type: ["object", "null"]
          additionalProperties: true
        createdAt:
          type: string
    ChatThreadDetail:
      type: object
      required: [id, messages]
      properties:
        id:
          type: string
        messages:
          type: array
          items:
            $ref: "#/components/schemas/ChatMessage"
    ChatTurnInput:
      type: object
      required: [userText, assistantPayload]
      properties:
        userText:
          type: string
        assistantPayload:
          type: object
          additionalProperties: true
...