# lib/api-zod/src/generated/types/authUser.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AuthUserRole } from './authUserRole';

export interface AuthUser {
  id: string;
  email: string;
  /** @nullable */
  name: string | null;
  role: AuthUserRole;
  /** Whether the user has accepted the current version of the legal documents. */
  legalAccepted: boolean;
}